+++ src/main/java/com/github/jlangch/venice/impl/SpecialForms.java
@@ -830,7 +830,7 @@ public class SpecialForms {
 		) {
 			private static final long serialVersionUID = -1;
 		};
-
+	
 	public static VncFunction macroexpand = 
 		new SpecialFormsDocFunction(
 				"macroexpand",